requirements.txt
fastapi
uvicorn
sqlalchemy
oracledb
python-dotenv
pydantic
pandas
google-genai
requests
streamlit

